+++ Source/NativeApi/NativeApi/Api/ListView/ListViewColumnEventData.cs
@@ -5,6 +5,6 @@ namespace NativeApi.Api
 {
     public class ListViewColumnEventData : NativeEventData
     {
-        public int columnIndex;
+        public long columnIndex;
     }
 }
